Fixes
git-svn-id: http://aptoide.org/repo/aptoide-client@1800 8d55eb69-6847-4b75-98a7-935b3913b0ec

+++ unstable/Aptoidev5/trunk/v5engine/src/main/java/cm/aptoide/ptdev/downloadmanager/DownloadInfo.java
@@ -329,12 +329,20 @@ public class DownloadInfo implements Runnable, Serializable {
     }
 
     public void remove() {
+
+
+        boolean isRemove = !getStatusState().getEnumState().equals(EnumState.COMPLETE);
+
         changeStatusState(new CompletedState(this));
 
         if (mFilesToDownload == null) return;
-        for (DownloadModel model : mFilesToDownload) {
-            new File(model.getDestination()).delete();
+
+        if(isRemove){
+            for (DownloadModel model : mFilesToDownload) {
+                new File(model.getDestination()).delete();
+            }
         }
+
         mProgress = 0;
         mSize = 0;
         mFilesToDownload.clear();
